refactor : 채팅 메시지 저장을 위한 mongoDB 설정 추가, JWT Bean 생성을 Core 모듈에서 처리하도록 변경 등
