+++ chat/build.gradle
@@ -18,6 +18,7 @@ configurations {
 }
 
 dependencies {
+    implementation project(path: ':core', configuration: 'default')
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
