+++ chat/src/main/resources/application.yml
@@ -1,4 +1,4 @@
 spring:
   profiles:
-    active: local
+    active: dev
     include: auth
