+++ chat/src/test/resources/application.yml
@@ -1,3 +1,4 @@
 spring:
   profiles:
     active: test
+    include: auth
